Heads and Tails buttons work!
Buttons can now be selected and WIN or LOSS will be displayed depending on coinflip. Additionally, not selecting Heads or Tails buttons will only show the coin flip result, not if the user won or not.

TODO: Choose a child to pick between Heads/Tails, store the coinflip result, and alternate so the other child gets a chance to pick between Heads/Tails.