Added more guidelines to fit different Android screen better, visual bug fixed, and more...
In activity_main.xml, I added guidelines to better fit small screens since the bottom buttons would at times be cut-off. Added guidelines leftCentreGuideline and rightCentreGuideline to ensure buttons are actually centred on every Android screen, instead of being 150dp from the left of the screen and the right side of the screen is cut-off for example.

In activity_coin_flip.xml, created leftCentreGuideline and rightCentreGuideline guidelines as described above. Also increased the font of FLIP button.

In activity_time_out.xml, added leftCentreGuideline and rightCentreGuideline guidelines. Added a horizontal guideline to fix the bug mentioned in the 'bug-report' channel on Discord.