Fixed FLIPCOIN button spam bug
Discovered that the coin flip animation would temporarily increase the coin size, pressing flip again would take that temporary size increase and increase it further.

Solution: Added a time delay that is delayed for the same duration that the coin animation is in effect. User cannot click flip again until the coin animation is over now.